requirements.txt
black
pylint
pyserial
